requirements.txt
numpy==1.19.5
pandas==1.2.4
matplotlib == 3.4.3
scikit-learn==0.24.2
joblib == 1.0.1
tpot == 0.11.7
pillow==8.2.0 
streamlit == 1.0.0



